Add files via upload
Updated the code so that the board is initialized with all lights turned off instead of filler. And re-worked the randomize_board function because the function was performing unnecessary checks on the lights. And added two functions, one called 'change_neighbors', the other named 'change_state'. Switched the values of the on and off variables because off was previously on, and on was previously off. Lastly, change_neighbors should take care of the illegal moves made by the computer when randomizing the board, thus making every possible board solvable.